# docs/openapi-user-api.yaml
openapi: 3.0.3
info:
  title: STEM Toys E-Commerce User Management API
  description: |
    Comprehensive User Management API for the STEM Toys Romanian E-Commerce platform.
    Features multi-tenant architecture, GDPR compliance, social authentication,
    advanced security, and Romanian market compliance.
  version: 1.0.0
  contact:
    name: STEM Toys Development Team
    email: dev@stemtoys.ro
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://stemtoys.ro/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT Bearer token for authenticated requests
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for administrative access
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: 'Invalid email address'
        code:
          type: string
          description: Error code for programmatic handling
          example: 'VALIDATION_ERROR'
...
paths:
...
  /consent:
    post:
      summary: Update user consent
      description: |
        Updates user consent for data processing according to GDPR requirements.
        All consent changes are logged for audit purposes.
      tags:
        - GDPR Compliance
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - consentType
                - consentGiven
              properties:
                consentType:
                  type: string
                  enum: [marketing, essential, analytics, personalization]
                  description: Type of consent being updated
                  example: 'marketing'
                consentGiven:
                  type: boolean
                  description: Whether consent is given or withdrawn
                  example: true
                validUntil:
                  type: string
                  format: date-time
                  description: Optional expiry date for time-limited consent
      responses:
        '200':
          description: Consent updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Consent updated successfully'
                  consentLog:
                    type: object
                    properties:
                      id:
                        type: string
                      consentType:
                        type: string
                      consentGiven:
                        type: boolean
                      timestamp:
                        type: string
                        format: date-time
        '400':
          description: Invalid consent data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /two-factor/setup:
    post:
      summary: Setup two-factor authentication
      description: Initiates 2FA setup and returns QR code for authenticator app
      tags:
        - Security
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 2FA setup initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: '2FA setup initiated'
                  secret:
                    type: string
                    description: TOTP secret for manual entry
                  qrCode:
                    type: string
                    description: Base64 encoded QR code image
                  backupCodes:
                    type: array
                    items:
                      type: string
                    description: Backup codes for account recovery
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /two-factor/verify:
    post:
      summary: Verify two-factor authentication
      description: Verifies 2FA token and enables 2FA for the account
      tags:
        - Security
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  description: 6-digit TOTP token from authenticator app
                  example: '123456'
      responses:
        '200':
          description: 2FA enabled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Two-factor authentication enabled successfully'
        '400':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /two-factor/disable:
    post:
      summary: Disable two-factor authentication
      description: Disables 2FA for the account (requires current password)
      tags:
        - Security
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  description: Current account password for verification
      responses:
        '200':
          description: 2FA disabled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Two-factor authentication disabled successfully'
        '400':
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'